% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataset.R
\name{createDataset}
\alias{createDataset}
\title{createDataset}
\usage{
createDataset(x, m, ...)
}
\arguments{
\item{x}{landslide points}

\item{m}{mask}

\item{...}{extra arguments}
}
\value{
Returns an object of class \code{sp}, which has landslide points and the same number of random points.
}
\description{
Creating dataset from the landslides points.
}
\details{
Returns an object of class \code{sp}, which has landslide points and the same number of random points.
}
\examples{
library(rgdal)

landslPoints<-readOGR(system.file("extdata/desliFixed.shp", package = "landsl"))

area<-readOGR(system.file("extdata/Region.shp", package = "landsl"))

createDataset(landslPoints, area)

}
