% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractAtr}
\alias{extractAtr}
\title{extractAtr}
\usage{
extractAtr(d, z, asp, slo, ...)
}
\arguments{
\item{d}{points dataset}

\item{z}{DEM elevation raster}

\item{asp}{aspect}

\item{slo}{slope}

\item{...}{extra arguments}
}
\value{
Returns an object of class \code{sp}, which has altitude, slope and aspect data.
}
\description{
Extract elevation, slope and aspect from rasater data.
}
\details{
Extract slope, elevations and aspect from raster data with that attribute base on the landslide points location.
}
\examples{

library(rgdal)
library(raster)


landslPoints<-readOGR(system.file("extdata/desliFixed.shp", package = "landsl"))

area<-readOGR(system.file("extdata/Region.shp", package = "landsl"))

demc<- raster(system.file("extdata/Mosaic_Clip.tif", package = "landsl"))

dem<-mask(demc, area)


slope <-slope(dem)

aspect<-aspect(dem)

ds <-createDataset(landslPoints, area)

extractAtr(ds, dem, aspect, slope)


}
