% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{probability}
\usage{
probability(d, DEM, asp, slo, ...)
}
\arguments{
\item{d}{dataset}

\item{DEM}{elevation raster}

\item{asp}{aspect rasater}

\item{slo}{slope raster}

\item{...}{extra arguments}
}
\value{
Returns an summary \code{data.frame} of INLA calculation; print dic and auc value and plot roc object.
}
\description{
Calculate the probability of landslides on the dataset points base on the informations they contain (slope, elevation and aspect).
}
\details{
This function calculates the probability
}
\examples{
library(rgdal)
library(raster)

landslPoints<-readOGR(system.file("extdata/desliFixed.shp", package = "landsl"))

area<-readOGR(system.file("extdata/Region.shp", package = "landsl"))

demc<- raster(system.file("extdata/Mosaic_Clip.tif", package = "landsl"))

dem<-mask(demc, area)
slope <-slope(dem)
aspect<-aspect(dem)
ds <-createDataset(landslPoints, area)
dsEx<-extractAtr(ds, dem, aspect, slope)
probability(dsEx, dem, aspect, slope)

}
