% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probMap}
\alias{probMap}
\title{probMap}
\usage{
probMap(sum, DEM, asp, slo, ...)
}
\arguments{
\item{sum}{summary from INLA}

\item{DEM}{elevation raster}

\item{asp}{aspect raster}

\item{slo}{slope raster}

\item{...}{extra arguments}
}
\value{
Returns an object of class \code{raster}, the probability of landslide in the study area.
}
\description{
Create a map to show probability of landslides on the dataset points base on the informations they contain (slope, elevation and aspect).
}
\details{
This function plots a landslide probability map
}
\examples{

library(rgdal)
library(raster)


landslPoints<-readOGR(system.file("extdata/desliFixed.shp", package = "landsl"))

area<-readOGR(system.file("extdata/Region.shp", package = "landsl"))

demc<- raster(system.file("extdata/Mosaic_Clip.tif", package = "landsl"))

dem<-mask(demc, area)

slope <-slope(dem)

aspect<-aspect(dem)

ds <-createDataset(landslPoints, area)

dsEx<-extractAtr(ds, dem, aspect, slope)

prob<-probability(dsEx, dem, aspect, slope)

probMap(prob, dem, aspect, slope)

}
